Fix duplicate operations
