+++ src/Tests/Scenarios.HandlerSelection/post_with_and_without_uri_var_as_string.cs
@@ -45,7 +45,7 @@ namespace Tests.Scenarios.HandlerSelection
         has
           .ResourcesNamed("example")
           .AtUri("/account/")
-          .And.AtUri("/account/?id={account_id")
+          .And.AtUri("/account/?id={account_id}")
           .HandledBy<CollectionHandler<string,string>>());
     }
